bookstore/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.6
django-celery-results==2.5.1
djangorestframework==3.15.1
exceptiongroup==1.2.1
iniconfig==2.0.0
kombu==5.3.7
packaging==24.0
pluggy==1.5.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pytest==8.2.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.4
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
